fix mistral moderation auto config
