+++ auto-configurations/models/spring-ai-autoconfigure-model-mistral-ai/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -15,3 +15,4 @@
 #
 org.springframework.ai.model.mistralai.autoconfigure.MistralAiChatAutoConfiguration
 org.springframework.ai.model.mistralai.autoconfigure.MistralAiEmbeddingAutoConfiguration
+org.springframework.ai.model.mistralai.autoconfigure.MistralAiModerationAutoConfiguration
